requirements.txt
discord.py>=2.3.0
flask>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
requests>=2.31.0
psycopg2-binary>=2.9.7
Pillow
google-generativeai
